# swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Rede Social",
    "description": "API de manipulação de dados do sistema de Rede Social.",
    "version": "1.0.0"
  },
  "host": "localhost:8686",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/auth": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "autenticação de usuário",
        "description": "Rota responsável por autenticar um usuário na Rede Social.",
        "operationId": "UserAuth",
        "parameters": [
          {
            "name": "input",
            "in": "body",
            "description": "E-mail e senha do usuário.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "timoteo@lob.com"
                },
                "senha": {
                  "type": "number",
                  "example": 4124241
                }
              },
              "required": [
                "email",
                "senha"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário autenticado!"
          },
          "400": {
            "description": "E-mail e/ou senha inválidos."
          }
        }
      }
    },
...
    "/group/members": {
      "put": {
        "tags": [
          "groups"
        ],
        "summary": "gerenciar membro de grupo",
        "description": "Rota responsável pela adição ou remoção de membro de um grupo da Rede Social.",
        "operationId": "ManageMember",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "description": "token gerado com JWT de autorização necessário para alterar as informações de um usuário",
            "in": "header",
            "type": "string",
            "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjYyYmRlNDMwOWM4N2VjM2ZlYTBmOWUwNCIsImVtYWlsIjoidGltb3Rlb0Bsb2IuY29tIiwiaWF0IjoxNjU2OTU2OTIwLCJleHAiOjE2NTcxMjk3MjB9.b5rzv5KKmEcwJ--A9N6RmWTw7KIE4GyN31wEPZXdm-Y"
          },
          {
            "name": "input",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id_group": {
                  "example": "62be02b2168cc021ff51479e"
                },
                "id_user": {
                  "example": "62bc8dc67ecd9c1757d023b2"
                },
                "action": {
                  "example": "add"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário adicionado ao grupo com sucesso!"
          },
          "401": {
            "description": "É preciso estar autenticado e ser um usuário administrador para adicionar ou remover usuários dos grupos."
          },
          "404": {
            "description": "Houve uma falha interna ao tentar realizar a ação..."
          }
        }
      }
    },
...
    "/posts/postsByUser": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "buscar publicações por ID de um usuário",
        "description": "Rota responsável por buscar as publicações pelo ID de um usuário da Rede Social.",
        "operationId": "GetPostsByUserID",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "description": "token gerado com JWT de autorização necessário para alterar as informações de um usuário",
            "in": "header",
            "type": "string",
            "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjYyYmRlNDMwOWM4N2VjM2ZlYTBmOWUwNCIsImVtYWlsIjoidGltb3Rlb0Bsb2IuY29tIiwiaWF0IjoxNjU2OTU2OTIwLCJleHAiOjE2NTcxMjk3MjB9.b5rzv5KKmEcwJ--A9N6RmWTw7KIE4GyN31wEPZXdm-Y"
          }
        ],
        "responses": {
          "200": {
            "description": "Busca de publicações do usuário realizada com sucesso!",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "_id": {
                    "example": "62bdc334eb5df8f45f8926a4"
                  },
                  "id_user": {
                    "example": "62bde4309c87ec3fea0f9e04"
                  },
                  "tema": { 
                    "example": "Férias e diversão"
                  },
                  "descricao": {
                    "example": "Olha que lugar bonito!"
                  },
                  "fotoPublicacao": {
                    "example": "http://fotospaisagens/ilha.png"
                  },
                  "curtidaDetalhe": {
                    "example": []
                  },
                  "comentarios": {
                    "example": []
                  },
                  "data": {
                    "example": "05/07/2022"
                  }
                }
              }
            }
          },
          "401": {
            "description": "É preciso estar logado para buscar as publicações!"
          },
          "404": {
            "description": "Houve uma falha interna ao buscar suas publicações..."
          }
        }
      }
    }
  }
}